Engin Demirog

+++ 04_Mathematics/FourOperations.cs
@@ -16,15 +16,15 @@ namespace _04_Mathematics
             Console.WriteLine("Subtraction Operation => {0} - {1} = {2}", number1, number2, result);
         }
 
-        public void Multiplication(int number1, int number2)
+        public void Multiplication(double number1, double number2)
         {
-            double result = (double)(number1 * number2);
+            double result = number1 * number2;
             Console.WriteLine("Multiplication Operation => {0} * {1} = {2}", number1, number2, result);
         }
 
-        public void Division(int number1, int number2)
+        public void Division(double number1, double number2)
         {
-            double result = (double)(number1 / number2);
+            double result = number1 / number2;
             Console.WriteLine("Division Operation => {0} / {1} = {2}", number1, number2, result);
         }
     }
